Trait aoc_nums::Int

pub trait Int: Clone + Copy + Debug + Default + Display + Eq + FromStr + Hash + Ord + IntConv + IntOps {
    type Signed: IntSigned;
    type Unsigned: IntUnsigned;

    const BITS: u32;
    const ZERO: Self;
    const ONE: Self;
    const TWO: Self;
    const THREE: Self;
    const FOUR: Self;
    const FIVE: Self;
    const SIX: Self;
    const SEVEN: Self;
    const EIGHT: Self;
    const NINE: Self;
    const MIN: Self;
    const MAX: Self;

    // Required methods
    fn unsigned_abs(self) -> Self::Unsigned;
    fn signum(self) -> Self::Signed;
    fn signed_diff(self, other: Self) -> NumResult<Self::Signed>;
    fn unsigned_diff(self, other: Self) -> NumResult<Self::Unsigned>;
    fn add_signed(self, other: Self::Signed) -> NumResult<Self>;
    fn sub_signed(self, other: Self::Signed) -> NumResult<Self>;
    fn gen_count_ones(self) -> u32;

    // Provided methods
    fn check_bit(self, bit: u32) -> bool { ... }
    fn bit_set_assign(&mut self, bit: u32) { ... }
    fn bit_clear_assign(&mut self, bit: u32) { ... }
    fn bound_start_assign(&mut self, other: Self) { ... }
    fn bound_end_assign(&mut self, other: Self) { ... }
    fn bounds_assign(&mut self, bounds: impl RangeBounds<Self>) { ... }
    fn lcm(num_0: Self, num_1: Self) -> Self { ... }
    fn gcd(num_0: Self, num_1: Self) -> Self { ... }

}

Required Associated Types

type Signed: IntSigned

type Unsigned: IntUnsigned

Required Associated Constants

const BITS: u32

const ZERO: Self

const ONE: Self

const TWO: Self

const THREE: Self

const FOUR: Self

const FIVE: Self

const SIX: Self

const SEVEN: Self

const EIGHT: Self

const NINE: Self

const MIN: Self

const MAX: Self

Required Methods

fn unsigned_abs(self) -> Self::Unsigned

fn signum(self) -> Self::Signed

fn signed_diff(self, other: Self) -> NumResult<Self::Signed>

Signed difference between two numbers

Errors

Returns Err (Overflow) if the result can’t be represented, ie if the difference is too high

fn unsigned_diff(self, other: Self) -> NumResult<Self::Unsigned>

Unsigned difference between two numbers

Errors

Returns Err (Overflow) if the result can’t be represented, ie if the second number is greater than the first

fn add_signed(self, other: Self::Signed) -> NumResult<Self>

Add a signed number

Errors

Returns Err (Overflow) if the result can’t be represented

fn sub_signed(self, other: Self::Signed) -> NumResult<Self>

Subtract a signed number

Errors

Returns Err (Overflow) if the result can’t be represented

fn gen_count_ones(self) -> u32

Generic wrapper for primitive count_ones function.

Provided Methods

fn check_bit(self, bit: u32) -> bool

fn bit_set_assign(&mut self, bit: u32)

fn bit_clear_assign(&mut self, bit: u32)

fn bound_start_assign(&mut self, other: Self)

fn bound_end_assign(&mut self, other: Self)

fn bounds_assign(&mut self, bounds: impl RangeBounds<Self>)

fn lcm(num_0: Self, num_1: Self) -> Self

fn gcd(num_0: Self, num_1: Self) -> Self