Trait aoc_nums::Int

pub trait Int: Clone + Copy + Debug + Default + Display + Eq + FromStr + Hash + Ord + IntOps + IntConv {
    type Signed: IntSigned;
    type Unsigned: IntUnsigned;

    const BITS: u32;
    const ZERO: Self;
    const ONE: Self;
    const TWO: Self;
    const THREE: Self;
    const FOUR: Self;
    const FIVE: Self;
    const SIX: Self;
    const SEVEN: Self;
    const EIGHT: Self;
    const NINE: Self;
    const MIN: Self;
    const MAX: Self;

    fn unsigned_abs(self) -> Self::Unsigned;
    fn signum(self) -> Self::Signed;
    fn signed_diff(self, other: Self) -> NumResult<Self::Signed>;
    fn unsigned_diff(self, other: Self) -> NumResult<Self::Unsigned>;
    fn add_signed(self, other: Self::Signed) -> NumResult<Self>;
    fn sub_signed(self, other: Self::Signed) -> NumResult<Self>;
    fn add_2(arg_0: Self, arg_1: Self) -> NumResult<Self>;
    fn div_2(arg_0: Self, arg_1: Self) -> NumResult<Self>;
    fn mul_2(arg_0: Self, arg_1: Self) -> NumResult<Self>;
    fn rem_2(arg_0: Self, arg_1: Self) -> NumResult<Self>;
    fn sub_2(arg_0: Self, arg_1: Self) -> NumResult<Self>;
    fn gen_count_ones(self) -> u32;

    fn add_3(arg_0: Self, arg_1: Self, arg_2: Self) -> NumResult<Self> { ... }
    fn add_4(arg_0: Self, arg_1: Self, arg_2: Self, arg_3: Self) -> NumResult<Self> { ... }
    fn mul_3(arg_0: Self, arg_1: Self, arg_2: Self) -> NumResult<Self> { ... }
    fn mul_4(arg_0: Self, arg_1: Self, arg_2: Self, arg_3: Self) -> NumResult<Self> { ... }
    fn check_bit(self, bit: u32) -> bool { ... }

}

Required Associated Types

type Signed: IntSigned

type Unsigned: IntUnsigned

Required Associated Constants

const BITS: u32

const ZERO: Self

const ONE: Self

const TWO: Self

const THREE: Self

const FOUR: Self

const FIVE: Self

const SIX: Self

const SEVEN: Self

const EIGHT: Self

const NINE: Self

const MIN: Self

const MAX: Self

Required Methods

fn unsigned_abs(self) -> Self::Unsigned

fn signum(self) -> Self::Signed

fn signed_diff(self, other: Self) -> NumResult<Self::Signed>

Signed difference between two numbers

Errors

Returns Err (Overflow) if the result can’t be represented, ie if the difference is too high

fn unsigned_diff(self, other: Self) -> NumResult<Self::Unsigned>

Unsigned difference between two numbers

Errors

Returns Err (Overflow) if the result can’t be represented, ie if the second number is greater than the first

fn add_signed(self, other: Self::Signed) -> NumResult<Self>

Add a signed number

Errors

Returns Err (Overflow) if the result can’t be represented

fn sub_signed(self, other: Self::Signed) -> NumResult<Self>

Subtract a signed number

Errors

Returns Err (Overflow) if the result can’t be represented

fn add_2(arg_0: Self, arg_1: Self) -> NumResult<Self>

Add two numbers together

Errors

Returns Err (Overflow) if the result can’t be represented

fn div_2(arg_0: Self, arg_1: Self) -> NumResult<Self>

Divide one number by another

Errors

Returns Err (Overflow) if the result can’t be represented

fn mul_2(arg_0: Self, arg_1: Self) -> NumResult<Self>

Multiply two numbers together

Errors

Returns Err (Overflow) if the result can’t be represented

fn rem_2(arg_0: Self, arg_1: Self) -> NumResult<Self>

Find the remainder from dividing one number by another

Errors

Returns Err (Overflow) if the result can’t be represented

fn sub_2(arg_0: Self, arg_1: Self) -> NumResult<Self>

Subtract one number from another

Errors

Returns Err (Overflow) if the result can’t be represented

fn gen_count_ones(self) -> u32

Provided Methods

fn add_3(arg_0: Self, arg_1: Self, arg_2: Self) -> NumResult<Self>

Add three numbers together

Errors

Returns Err (Overflow) if the result can’t be represented

fn add_4(arg_0: Self, arg_1: Self, arg_2: Self, arg_3: Self) -> NumResult<Self>

Add four numbers together

Errors

Returns Err (Overflow) if the result can’t be represented

fn mul_3(arg_0: Self, arg_1: Self, arg_2: Self) -> NumResult<Self>

Multiply three numbers together

Errors

Returns Err (Overflow) if the result can’t be represented

fn mul_4(arg_0: Self, arg_1: Self, arg_2: Self, arg_3: Self) -> NumResult<Self>

Multiply four numbers together

Errors

Returns Err (Overflow) if the result can’t be represented

fn check_bit(self, bit: u32) -> bool